move settings UI from login to dedicated file better code and RAM management, pylint
